fix bugs in views and add tests

+++ src/Views/SearchViewPort.cpp
@@ -3,7 +3,7 @@
 void SearchViewport::updateHorizontal(const SearchSession &session, const TextLayout &textLayout, const SearchLayout &layout)
 {
     uint32_t cursorTextWidth = textLayout.width(session.getQuery().substr(0, session.getCursor()));
-    mHorizontal.visibleWidth = static_cast<int>(layout.queryBox.w) - static_cast<int>(layout.textPadding * 2);
+    mHorizontal.visibleWidth = layout.queryBox.w - layout.textPadding * 2;
     mHorizontal.ensureVisible(static_cast<int>(cursorTextWidth), 2);
 }
